refactor: 디렉터리 이름 정리
