+++ src/main/java/common/GameStatus.java
@@ -1,4 +1,4 @@
-package domain.game;
+package common;
 
 
 public enum GameStatus {
