+++ src/main/java/db/DbConnectionFactory.java
@@ -1,4 +1,4 @@
-package infra.db;
+package db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
