+++ src/main/java/service/JanggiGameSetupService.java
@@ -1,4 +1,4 @@
-package application;
+package service;
 
 import static domain.player.Team.CHO;
 import static domain.player.Team.HAN;
@@ -10,7 +10,7 @@ import dao.GameRoom;
 import dao.PlayerNames;
 import domain.board.Board;
 import domain.board.Formation;
-import domain.game.GameStatus;
+import common.GameStatus;
 import domain.manager.GameManager;
 import domain.piece.BasicPiece;
 import domain.player.Name;
