+++ src/test/java/dao/BoardRepositoryTest.java
@@ -12,8 +12,8 @@ import domain.board.Formation;
 import domain.piece.BasicPiece;
 import domain.piece.None;
 import domain.position.Position;
-import infra.db.DbBootstrap;
-import infra.db.DbConnectionFactory;
+import db.DbBootstrap;
+import db.DbConnectionFactory;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
